Rename RoadBook MenuItem & move Picture tests
