+++ app/src/main/res/values/strings.xml
@@ -15,7 +15,7 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
 
-    <string name="menu_roadbook">Home</string>
+    <string name="menu_roadbook">RoadBook</string>
     <string name="menu_directory">Directory</string>
     <string name="menu_picture">Picture</string>
     <string name="fab_add_description">add_floating_button</string>
